+++ firmware/main/network/flex/SampleRateConverter.h
@@ -34,7 +34,7 @@ extern "C" {
    The Flex uses floating point numbers while Codec2 uses shorts, so we
    also have code to convert between those formats. */
 void           fdmdv_8_to_24(float out24k[], short in8k[], int n);
-void           fdmdv_24_to_8(short out8k[], float in24k[], int n);
+void           fdmdv_24_to_8(short out8k[], short in24k[], int n);
 
 #ifdef __cplusplus
 }
